app/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
numpy==2.2.3
requests==2.32.3
streamlit==1.43.1
python-multipart==0.0.20
